backend/store/requirements.txt
django == 5.1.4
django-cors-headers == 4.3.1
django-allauth == 0.61.1
dj-rest-auth == 5.0.2
django-filter == 23.5
python-dotenv == 1.0.1
python-decouple == 3.8
django-import-export == 3.3.6
pillow == 11.1.0